# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Lok Lingu API — voice-controlled language learning arcade runner
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserInput {
  /**
     * @minLength 1
     * @maxLength 32
     */
  username: string;
}

export interface User {
  id: number;
  username: string;
  tokenBalance?: number;
  createdAt: string;
}

export interface ScoreInput {
  userId: number;
  language: string;
  category: string;
  /** @minimum 0 */
  count: number;
}

export interface Score {
  id: number;
  userId: number;
  language: string;
  category: string;
  count: number;
  createdAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  count: number;
  language: string;
  category: string;
  createdAt?: string;
}

export interface LeaderboardSummary {
  totalPlayers: number;
  totalGames: number;
  allTimeHighCount: number;
  /** @nullable */
  topPlayer: string | null;
}

export interface Word {
  word: string;
  translation: string;
  index: number;
  /** @nullable */
  pronunciation?: string | null;
}

export interface Language {
  code: string;
  name: string;
  categories: string[];
}

export interface GlobalStats {
  totalPlayers: number;
  totalGames: number;
  totalWordsSpoken: number;
  allTimeHighCount: number;
}

export interface UserStats {
  userId: number;
  username: string;
  totalGames: number;
  bestCount: number;
  totalWordsSpoken: number;
  /** @nullable */
  favoriteLanguage: string | null;
  /** @nullable */
  favoriteCategory: string | null;
  personalBests?: Score[];
}

export type GetLeaderboardParams = {
language?: string;
category?: string;
limit?: number;
};