# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Lok Lingu API — voice-controlled language learning arcade runner
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  count: number;
  language: string;
  category: string;
  createdAt?: Date;
}