# lib/api-zod/src/generated/types/score.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Lok Lingu API — voice-controlled language learning arcade runner
 * OpenAPI spec version: 0.1.0
 */

export interface Score {
  id: number;
  userId: number;
  language: string;
  category: string;
  count: number;
  createdAt: Date;
}