# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Lok Lingu API — voice-controlled language learning arcade runner
 * OpenAPI spec version: 0.1.0
 */
import type { Score } from './score';

export interface UserStats {
  userId: number;
  username: string;
  totalGames: number;
  bestCount: number;
  totalWordsSpoken: number;
  /** @nullable */
  favoriteLanguage: string | null;
  /** @nullable */
  favoriteCategory: string | null;
  personalBests?: Score[];
}